+++ packages/agent/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Fixed `PI_DIALECT=minimax` being ignored by the owned tool-calling env selector. ([#2759](https://github.com/can1357/oh-my-pi/issues/2759))
 
+### Added
+
+- Added agent-loop deadline support for graceful wall-clock session stops.
+
 ## [16.0.1] - 2026-06-15
 
 ### Fixed